exported and moved authOptions
