+++ projects/angular-auth-oidc-client/src/lib/auth-options.ts
