+++ projects/angular-auth-oidc-client/src/lib/login/standard/standard-login.service-mock.ts
@@ -1,4 +1,4 @@
-import { AuthOptions } from '../auth-options';
+import { AuthOptions } from '../../auth-options';
 
 export class StandardLoginServiceMock {
   loginStandard(authOptions?: AuthOptions) {}
